Flesh out mounting and components
Components -> A module that returns a router

Middleware -> similar to a component returns a connect style middleware
function. Which can be built off a top level router.

Logger -> A built in top level functionality which supports being
overriden

Router -> Top level function that returns a chainable router object
which in turn can be passed back as a component.